requirements.txt
asn1crypto==0.24.0
astroid==2.3.3
certifi==2019.11.28
chardet==3.0.4
command-not-found==0.3
cryptography==2.1.4
cupshelpers==1.0
distro-info===0.18ubuntu0.18.04.1
httplib2==0.9.2
idna==2.9
isodate==0.6.0
isort==4.3.21
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
lazy-object-proxy==1.4.3
mccabe==0.6.1
mypy-extensions==0.4.3
netifaces==0.10.4
olefile==0.45.1
pexpect==4.2.1
Pillow==7.0.0
progressbar2==3.50.0
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
pygobject==3.26.1
pylint==2.4.4
pyparsing==2.4.6
python-apt==1.6.5+ubuntu0.2
python-utils==2.4.0
pyxdg==0.25
PyYAML==3.12
qwikidata==0.4.0
rdflib==4.2.2
reportlab==3.4.0
requests==2.23.0
requests-unixsocket==0.1.5
SecretStorage==2.3.1
sip==4.19.21
six==1.14.0
SPARQLWrapper==1.8.5
staticmap==0.5.4
typed-ast==1.4.1
ubuntu-drivers-common==0.0.0
ufw==0.36
urllib3==1.25.8
wrapt==1.11.2
xkit==0.0.0
